+++ src/available-cred.ts
